refactor(demos): optimize data movement and thread management
Improve efficiency and resource handling in demonstration files:
- Use `std::move` for transferring large vectors (sequence and pose bytes) to the blackboard.
- Refactor `FibonacciActionServer` to use a managed `GoalThread` structure with `std::atomic` flags for cleaner thread lifecycle management and non-blocking cleanup.
- Update `parameters_demo` to use `int64_t` for counter comparisons to ensure type consistency.

+++ yasmin_demos/src/parameters_demo.cpp
@@ -76,7 +76,7 @@ public:
     YASMIN_LOG_INFO("Executing state FOO");
     std::this_thread::sleep_for(std::chrono::seconds(3));
 
-    if (this->counter < blackboard->get<int>("max_counter")) {
+    if (this->counter < blackboard->get<int64_t>("max_counter")) {
       this->counter += 1;
       blackboard->set<std::string>("foo_str",
                                    blackboard->get<std::string>("counter_str") +
